Fix for Windows ResEditor project and crash at exporting
Former-commit-id: d65a05abcfc904c6556a4864092db4e0b6ba46d4

+++ Sources/Internal/Scene3D/LandscapeNode.cpp
@@ -1436,7 +1436,7 @@ Texture * LandscapeNode::CreateFullTiledTexture()
     
     Texture *fullTiled = Texture::CreateFBO(TEXTURE_TILE_FULL_SIZE, TEXTURE_TILE_FULL_SIZE, FORMAT_RGBA8888, Texture::DEPTH_NONE);
     RenderManager::Instance()->SetRenderTarget(fullTiled);
-    RenderManager::Instance()->SetViewport(Rect(0.f, 0.f, fullTiled->GetWidth(), fullTiled->GetHeight()), true);
+    RenderManager::Instance()->SetViewport(Rect(0.f, 0.f, (float32)fullTiled->GetWidth(), (float32)fullTiled->GetHeight()), true);
 
 
 	RenderManager::Instance()->ClearWithColor(1.f, 1.f, 1.f, 1.f);
